final_project/requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
datasets==3.1.0
Deprecated==1.2.15
dill==0.3.8
evaluate==0.4.3
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.1
frozenlist==1.5.0
fsspec==2024.9.0
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
neural_compressor==3.1.1
numpy==1.23.5
onnx==1.17.0
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
optimum==1.23.3
optimum-intel==1.20.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
prettytable==3.12.0
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.1.0
pycocotools==2.0.8
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
schema==0.7.7
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
